Fixes by Volker Christian
git-svn-id: svn://svn.code.sf.net/p/synce/code/trunk@1784 599ab11a-0111-0410-8a50-d7a5b0953570

+++ rra/src/rra-appointment-to-vevent.c
@@ -51,7 +51,7 @@ int main(int argc, char** argv)
       else
       {
         fprintf(stderr, "%s: Only read %i bytes from time zone information file '%s': %s\n", 
-            argv[0], bytes_read, argv[2], strerror(errno));
+            argv[0], (int) bytes_read, argv[2], strerror(errno));
       }
 
       fclose(file);
